+++ src/api.ts
@@ -1,22 +1,31 @@
 import urlcat from 'urlcat'
 import { apiUrl, baseUrl } from './constants'
-import { getChatIdFromUrl, getConversationFromSharePage, isSharePage } from './page'
+import { getChatIdFromUrl, getConversationFromSharePage, getPageAccessToken, isSharePage } from './page'
 import { blobToDataURL } from './utils/dom'
+import { memorize } from './utils/memorize'
 
 interface ApiSession {
     accessToken: string
+    authProvider: string
     expires: string
     user: {
         email: string
         groups: string[]
+        // token's issued_at timestamp
+        iat: number
         id: string
+        // token's expiration timestamp
+        idp: string
         image: string
+        intercom_hash: string
+        // whether the user has multi-factor authentication enabled
+        mfa: boolean
         name: string
         picture: string
     }
 }
 
-type ModelSlug = 'text-davinci-002-render-sha' | 'text-davinci-002-render-paid' | 'text-davinci-002-browse' | 'gpt-4' | 'gpt-4-browsing'
+type ModelSlug = 'text-davinci-002-render-sha' | 'text-davinci-002-render-paid' | 'text-davinci-002-browse' | 'gpt-4' | 'gpt-4-browsing' | 'gpt-4o'
 
 export interface Citation {
     start_ix: number
@@ -137,6 +146,9 @@ export interface ConversationNodeMessage {
         // multi-modal input
         content_type: 'multimodal_text'
         parts: Array<MultiModalInputImage | string>
+    } | {
+        content_type: 'model_editable_context'
+        model_set_context: string
     }
     create_time?: number
     update_time?: number
@@ -189,6 +201,46 @@ export interface ApiConversations {
     total: number
 }
 
+interface ApiAccountsCheckAccountDetail {
+    account_user_role: 'account-owner' | string
+    account_user_id: string | null
+    processor: Record<string, boolean>
+    account_id: string | null
+    organization_id?: string | null
+    is_most_recent_expired_subscription_gratis: boolean
+    has_previously_paid_subscription: boolean
+    name?: string | null
+    profile_picture_id?: string | null
+    profile_picture_url?: string | null
+    structure: 'workspace' | 'personal'
+    plan_type: 'team' | 'free'
+    is_deactivated: boolean
+    promo_data: Record<string, unknown>
+}
+
+interface ApiAccountsCheckEntitlement {
+    subscription_id?: string | null
+    has_active_subscription?: boolean
+    subscription_plan?: 'chatgptteamplan' | 'chatgptplusplan'
+    expires_at?: string | null
+    billing_period?: 'monthly' | string | null
+}
+
+interface ApiAccountsCheckAccount {
+    account: ApiAccountsCheckAccountDetail
+    features: string[]
+    entitlement: ApiAccountsCheckEntitlement
+    last_active_subscription?: Record<string, unknown> | null
+    is_eligible_for_yearly_plus_subscription: boolean
+}
+
+interface ApiAccountsCheck {
+    accounts: {
+        [key: string]: ApiAccountsCheckAccount
+    }
+    account_ordering: string[]
+}
+
 type ApiFileDownload = {
     status: 'success'
     /** signed download url */
@@ -203,10 +255,34 @@ type ApiFileDownload = {
     error_message: string | null
 }
 
+// eslint-disable-next-line no-restricted-syntax
+const enum ChatGPTCookie {
+    AgeVerification = 'oai-av-seen',
+    AllowNonessential = 'oai-allow-ne',
+    DeviceId = 'oai-did',
+    DomainMigrationSourceCompleted = 'oai-dm-src-c-240329',
+    DomainMigrationTargetCompleted = 'oai-dm-tgt-c-240329',
+    HasClickedOnTryItFirstLink = 'oai-tif-20240402',
+    HasLoggedInBefore = 'oai-hlib',
+    HideLoggedOutBanner = 'hide-logged-out-banner',
+    IntercomDeviceIdDev = 'intercom-device-id-izw1u7l7',
+    IntercomDeviceIdProd = 'intercom-device-id-dgkjq2bp',
+    IpOverride = 'oai-ip-country',
+    IsEmployee = '_oaiauth',
+    IsPaidUser = '_puid',
+    LastLocation = 'oai-ll',
+    SegmentUserId = 'ajs_user_id',
+    SegmentUserTraits = 'ajs_user_traits',
+    ShowPaymentModal = 'ui-show-payment-modal',
+    TempEnableUnauthedCompliance = 'temp-oai-compliance',
+    Workspace = '_account',
+}
+
 const sessionApi = urlcat(baseUrl, '/api/auth/session')
 const conversationApi = (id: string) => urlcat(apiUrl, '/conversation/:id', { id })
 const conversationsApi = (offset: number, limit: number) => urlcat(apiUrl, '/conversations', { offset, limit })
 const fileDownloadApi = (id: string) => urlcat(apiUrl, '/files/:id/download', { id })
+const accountsCheckApi = urlcat(apiUrl, '/accounts/check/v4-2023-04-27')
 
 export async function getCurrentChatId(): Promise<string> {
     if (isSharePage()) {
@@ -346,12 +422,14 @@ export async function deleteConversation(chatId: string): Promise<boolean> {
 
 async function fetchApi<T>(url: string, options?: RequestInit): Promise<T> {
     const accessToken = await getAccessToken()
+    const accountId = await getTeamAccountId()
 
     const response = await fetch(url, {
         ...options,
         headers: {
             'Authorization': `Bearer ${accessToken}`,
             'X-Authorization': `Bearer ${accessToken}`,
+            ...(accountId ? { 'Chatgpt-Account-Id': accountId } : {}),
             ...options?.headers,
         },
     })
@@ -361,20 +439,54 @@ async function fetchApi<T>(url: string, options?: RequestInit): Promise<T> {
     return response.json()
 }
 
-async function getAccessToken(): Promise<string> {
-    const session = await fetchSession()
-    return session.accessToken
-}
-
-let session: ApiSession | null = null
-async function fetchSession(): Promise<ApiSession> {
-    if (session) return session
+async function _fetchSession(): Promise<ApiSession> {
     const response = await fetch(sessionApi)
     if (!response.ok) {
         throw new Error(response.statusText)
     }
-    session = await response.json()
-    return session!
+    return response.json()
+}
+
+const fetchSession = memorize(_fetchSession)
+
+async function getAccessToken(): Promise<string> {
+    const pageAccessToken = getPageAccessToken()
+    if (pageAccessToken) return pageAccessToken
+
+    const session = await fetchSession()
+    return session.accessToken
+}
+
+async function _fetchAccountsCheck(): Promise<ApiAccountsCheck> {
+    const accessToken = await getAccessToken()
+
+    const response = await fetch(accountsCheckApi, {
+        headers: {
+            'Authorization': `Bearer ${accessToken}`,
+            'X-Authorization': `Bearer ${accessToken}`,
+        },
+    })
+    if (!response.ok) {
+        throw new Error(response.statusText)
+    }
+    return response.json()
+}
+
+const fetchAccountsCheck = memorize(_fetchAccountsCheck)
+
+const getCookie = (key: string) => document.cookie.match(`(^|;)\\s*${key}\\s*=\\s*([^;]+)`)?.pop() || ''
+
+export async function getTeamAccountId(): Promise<string | null> {
+    const accountsCheck = await fetchAccountsCheck()
+    const workspaceId = getCookie(ChatGPTCookie.Workspace)
+    if (workspaceId) {
+        const account = accountsCheck.accounts[workspaceId]
+        if (account) {
+            return account.account.account_id
+        }
+    }
+
+    return null
 }
 
 export interface ConversationResult {
@@ -393,6 +505,7 @@ const ModelMapping: { [key in ModelSlug]: string } & { [key: string]: string } =
     'text-davinci-002-browse': 'GPT-3.5',
     'gpt-4': 'GPT-4',
     'gpt-4-browsing': 'GPT-4 (Browser)',
+    'gpt-4o': 'GPT-4o',
 
     // fuzzy matching
     'text-davinci-002': 'GPT-3.5',
@@ -458,7 +571,12 @@ function extractConversationResult(conversationMapping: Record<string, Conversat
             break // Stop at root message.
         }
 
-        if (node.message?.author.role !== 'system') { // Skip system messages
+        if (
+            // Skip system messages
+            node.message?.author.role !== 'system'
+            // Skip model memory context
+            && node.message?.content.content_type !== 'model_editable_context'
+        ) {
             result.unshift(node)
         }