+++ vite.config.ts
@@ -30,8 +30,40 @@ export default defineConfig({
                     'zh-TW': packageJson['description:zh-TW'],
                 },
                 'license': packageJson.license,
+                // 需要对其它网站生效的时候改为'*://*/*'
                 'match': [
-                    '*://*/*',
+                    'https://chat.openai.com/',
+                    // support https://chat.openai.com/?model={model}
+                    'https://chat.openai.com/?model=*',
+                    // support https://chat.openai.com/c/123456789
+                    'https://chat.openai.com/c/*',
+                    // support https://chat.openai.com/g/g-123456789
+                    'https://chat.openai.com/g/*',
+                    // support https://chat.openai.com/gpts/
+                    'https://chat.openai.com/gpts',
+                    'https://chat.openai.com/gpts/*',
+                    // support https://chat.openai.com/share/123456789
+                    'https://chat.openai.com/share/*',
+                    // support https://chat.openai.com/share/123456789/continue
+                    'https://chat.openai.com/share/*/continue',
+
+                    'https://chatgpt.com/',
+                    'https://chatgpt.com/?model=*',
+                    'https://chatgpt.com/c/*',
+                    'https://chatgpt.com/g/*',
+                    'https://chatgpt.com/gpts',
+                    'https://chatgpt.com/gpts/*',
+                    'https://chatgpt.com/share/*',
+                    'https://chatgpt.com/share/*/continue',
+
+                    'https://new.oaifree.com/',
+                    'https://new.oaifree.com/?model=*',
+                    'https://new.oaifree.com/c/*',
+                    'https://new.oaifree.com/g/*',
+                    'https://new.oaifree.com/gpts',
+                    'https://new.oaifree.com/gpts/*',
+                    'https://new.oaifree.com/share/*',
+                    'https://new.oaifree.com/share/*/continue',
                 ],
                 'icon': 'https://chat.openai.com/favicon.ico',
                 'run-at': 'document-end',
@@ -42,6 +74,17 @@ export default defineConfig({
                     ['jszip', cdn.jsdelivr('JSZip', 'dist/jszip.min.js')],
                     ['html2canvas', cdn.jsdelivr('html2canvas', 'dist/html2canvas.min.js')],
                 ],
+                cssSideEffects() {
+                    return (e) => {
+                        const o = document.createElement('style')
+                        o.textContent = e
+                        document.head.append(o)
+                        setInterval(() => {
+                            if (o.isConnected) return
+                            document.head.append(o)
+                        }, 300)
+                    }
+                },
             },
             server: {
                 open: true,